requirements.txt
Django
psycopg2-binary
requests
Pillow
PyJWT
apscheduler
pytz
pydantic
ipython
